+++ __mocks__/api_data.json
@@ -22,6 +22,31 @@
       "email_duplicate_error": true,
       "message": true,
       "success": false
-    }
+    },
+    "profiles":[
+      {
+        "id": "1",
+        "registration": "123456",
+        "sector_id": "1",
+        "hospital_id": "1",
+        "user_id": "1"
+      }
+    ],
+    "scheduleSector":[
+      {
+        "start_time": "2018-06-24T08:00",
+        "end_time": "2018-07-25T08:00",
+        "profile_id": "1",
+        "sector_id": "1"
+      }
+    ],
+    "scheduleProfiles":[
+      {
+      "start_time": "'2018-06-24T08:00'",
+      "end_time": "2018-07-25T08:00",
+      "profile_id": "1",
+      "sector_id": "1"
+      }
+    ]
   }
 }
